+++ src/card/sets/y2017/mod.rs
@@ -1,6 +1,7 @@
 //! Card sets first released in 2017.
 
 pub(super) mod aether_revolt;
+pub(super) mod amonkhet;
 pub(super) mod ixalan;
 
-pub(super) use super::{CardRecord, PrintingRecord};
+pub(super) use super::{CardRecord, PrintingAnchor, PrintingRecord};